I added a command line option for Verbosity Level, this solves the compile-time error that complained about the variable verbosity not being defined.
I fixed the shading issue with observed vs. hidden nodes.  This required a deep
copy of an object newRvi, instead of a copy of the pointer because once the data
was set up, the FileParser object (which contained this newRvi object) was
freed.  This also ment that I needed to delete the object in the VizNode
destructor.

I increased the size of the Arc control points as I wasn't able to see them in
the default view.

Finally, I changed the formatting (whitespace, tabs, etc) to be more uniform and
pleasing to me.  If this gives you a problem with cvs diff use diff -b [which
ignores whitespace differences].