I've changed the uncommented the default pen styles for the various types of lines (ie the grid, edges between nodes with various relationships, etc), so that there are dotted and dashed lines by default.  I disabled zoom in the print preview by subclassing the wxPrintPreview class and making the zoom function do nothing.  I look into this problem more and hopefully allow for zoom in the print preview in the future.  Currently printing works for all of the example input files I have (even with dotted and dashed lines).
I changed some formatting (use cvs diff -b if you want to ignore whitespace
differences).

I noticed that the menu item that allowed for a user to change the canvas
height actually changed the width so I fixed that.

I made the copy frame layout show a user selectable list of frames to apply the
layout to.  This allows for multiple frames to be affected by the layout.
Eventually I'd like to make a custom dialog that consolidates this whole
procedure and uses one GUI that allows for selection of the source and
destionation frame(s).

I changed 2 comments that I should have changed on the last commit, they
refered to a pointer into the FileParser's data, we're not copying the data so
it is no longer a Pointer into the FileParser.