Initial version of the class GMTK_Tie which handles the clustering, tying, and untying of model parameters. It also loads and parses the user-supplied command file. This new class was made to simplify the interface to the parameter classes (GMTK_Tie is a friend of all the appropriate classes).
In this version, only user-specified ties/unties and data-driven,
bottom up agglomerative clustering (followed by within-cluster tying)
are implemented, and for a limited set of parameter types (Mixture,
MeanVector). The framework is in place to extend to all other
parameter types later.

A variety of dissimilarity measures, cluster size calculation methods,
outlier removal crietria, etc are implemented or at least have
placeholders in the code. This is easily extensible.

Mixtures are currently tied at the Component/dense1DPMF level but this
will probably change to the Mixture level in a later version.